Move subviews and action handlers to UIToolbar view subclass from view controller to be able to reuse it for SwiftUI layout
