+++ catowser/catowser/WebBrowserToolbar/WebBrowserToolbarController.swift
@@ -18,158 +18,47 @@ import UIKit
 import CoreBrowser
 
 final class WebBrowserToolbarController<C: Navigating>: BaseViewController where C.R == ToolbarRoute {
-    
     private weak var coordinator: C?
-
-    init(_ coordinator: C,
-         _ downloadDelegate: DownloadPanelPresenter?,
-         _ globalSettingsDelegate: GlobalMenuDelegate?) {
-        self.coordinator = coordinator
-        downloadPanelDelegate = downloadDelegate
-        self.globalSettingsDelegate = globalSettingsDelegate
-        downloadsViewHidden = true
-        enableDownloadsButton = false
-        super.init(nibName: nil, bundle: nil)
-        TabsListManager.shared.attach(counterView)
-        TabsListManager.shared.attach(self)
-    }
-    
-    /// Site navigation delegate
+    /// download panel delegate
+    private weak var downloadPanelDelegate: DownloadPanelPresenter?
+    /// web view interface
     private weak var webViewInterface: WebViewNavigatable? {
         didSet {
             guard webViewInterface != nil else {
-                backButton.isEnabled = false
-                forwardButton.isEnabled = false
-                reloadButton.isEnabled = false
-                downloadsViewHidden = true
-                enableDownloadsButton = false
+                toolbarView.state = .nothingToNavigate
                 return
             }
-
-            reloadNavigationElements(true)
+            toolbarView.webViewInterface = webViewInterface
+            toolbarView.state = .readyForNavigation
         }
     }
     
-    private weak var downloadPanelDelegate: DownloadPanelPresenter?
-    private weak var globalSettingsDelegate: GlobalMenuDelegate?
+    private let toolbarView: WebBrowserToolbarView
 
-    fileprivate var downloadsViewHidden: Bool = true {
-        didSet {
-            animateDownloadsArrow(down: !downloadsViewHidden)
-        }
+    init(_ coordinator: C,
+         _ downloadPanelDelegate: DownloadPanelPresenter?,
+         _ globalSettingsDelegate: GlobalMenuDelegate?) {
+        self.coordinator = coordinator
+        self.downloadPanelDelegate = downloadPanelDelegate
+        toolbarView = WebBrowserToolbarView(frame: .zero)
+        toolbarView.globalSettingsDelegate = globalSettingsDelegate
+        super.init(nibName: nil, bundle: nil)
     }
     
-    fileprivate var enableDownloadsButton: Bool {
-        didSet {
-            downloadLinksButton.isEnabled = enableDownloadsButton
-            downloadsView.alpha = enableDownloadsButton ? 1.0 : 0.3
-        }
-    }
-    
-    private lazy var standardToolbarButtons: [UIBarButtonItem] = {
-        var barItems = [UIBarButtonItem]()
-        barItems.append(backButton)
-        let space1 = UIBarButtonItem(barButtonSystemItem: .flexibleSpace, target: nil, action: nil)
-        barItems.append(space1)
-        barItems.append(forwardButton)
-        let space2 = UIBarButtonItem(barButtonSystemItem: .flexibleSpace, target: nil, action: nil)
-        barItems.append(space2)
-        barItems.append(reloadButton)
-        let space3 = UIBarButtonItem(barButtonSystemItem: .flexibleSpace, target: nil, action: nil)
-        barItems.append(space3)
-        barItems.append(openedTabsButton)
-        return barItems
-    }()
-
-    private lazy var toolbarView: CottonToolbarView = {
-        let toolbar = CottonToolbarView(frame: .zero)
-        ThemeProvider.shared.setup(toolbar)
-        toolbar.counterView = counterView
-        toolbar.downloadsView = downloadsView
-        toolbar.setItems(standardToolbarButtons, animated: false)
-        return toolbar
-    }()
-    
-    private lazy var backButton: UIBarButtonItem = {
-        let img = UIImage(named: "nav-back")
-        let back = #selector(WebBrowserToolbarController.handleBackPressed)
-        let btn = UIBarButtonItem(image: img, style: .plain, target: self, action: back)
-        return btn
-    }()
-    
-    private lazy var forwardButton: UIBarButtonItem = {
-        let img = UIImage(named: "nav-forward")
-        let forward = #selector(WebBrowserToolbarController.handleForwardPressed)
-        let btn = UIBarButtonItem(image: img, style: .plain, target: self, action: forward)
-        return btn
-    }()
-    
-    private lazy var reloadButton: UIBarButtonItem = {
-        let img = UIImage(named: "nav-refresh")
-        let reload = #selector(WebBrowserToolbarController.handleReloadPressed)
-        let btn = UIBarButtonItem(image: img, style: .plain, target: self, action: reload)
-        return btn
-    }()
-    
-    private lazy var actionsButton: UIBarButtonItem = {
-        let btn: UIBarButtonItem
-        let actions = #selector(WebBrowserToolbarController.handleActionsPressed)
-        if #available(iOS 13.0, *) {
-            if let systemImage = UIImage.arropUp {
-                btn = .init(image: systemImage, style: .plain, target: self, action: actions)
-            } else {
-                btn = .init(barButtonSystemItem: .action, target: self, action: actions)
-            }
-        } else {
-            btn = .init(barButtonSystemItem: .action, target: self, action: actions)
-        }
-        return btn
-    }()
-
-    private let counterView: CounterView = CounterView(frame: .zero)
-    
-    private lazy var openedTabsButton: UIBarButtonItem = {
-        counterView.digit = TabsListManager.shared.tabsCount
-        // Can't use simple bar button with text because it positioned incorrectly
-        let btn = UIBarButtonItem(customView: counterView)
-        return btn
-    }()
-
-    private lazy var downloadsView: UIImageView = {
-        let img = UIImage(named: "nav-downloads")
-        let imgView = UIImageView(image: img)
-        return imgView
-    }()
-    
-    private lazy var downloadLinksButton: UIBarButtonItem = {
-        let downloads = #selector(WebBrowserToolbarController.handleDownloadsPressed)
-        let btn = UIBarButtonItem(customView: downloadsView)
-        btn.target = self
-        btn.action = downloads
-        return btn
-    }()
-
-    deinit {
-        TabsListManager.shared.detach(self)
-        TabsListManager.shared.detach(counterView)
-    }
-
     required init?(coder aDecoder: NSCoder) {
         fatalError("init(coder:) has not been implemented")
+    }
+    
+    deinit {
     }
 
     override func loadView() {
         view = toolbarView
+        ThemeProvider.shared.setup(toolbarView)
     }
     
     override func viewDidLoad() {
         super.viewDidLoad()
-
-        // disabled after `init` because no web view is present
-        backButton.isEnabled = false
-        forwardButton.isEnabled = false
-        reloadButton.isEnabled = false
-        downloadLinksButton.isEnabled = false
     }
 
     override func touchesBegan(_ touches: Set<UITouch>, with event: UIEvent?) {
@@ -177,10 +66,10 @@ final class WebBrowserToolbarController<C: Navigating>: BaseViewController where
         // for strange reason it can't recognize gesture recognizers or
         // even target-action for this specific view
         for touch in touches {
-            if touch.view == counterView {
+            if touch.view == toolbarView.counterView {
                 handleShowOpenedTabsPressed()
                 break
-            } else if touch.view == downloadsView && enableDownloadsButton {
+            } else if touch.view == toolbarView.downloadsView {
                 handleDownloadsPressed()
                 break
             }
@@ -189,80 +78,41 @@ final class WebBrowserToolbarController<C: Navigating>: BaseViewController where
     
     // MARK: - private functions
     
-    private func animateDownloadsArrow(down: Bool) {
-        let rotate = UIViewPropertyAnimator(duration: 0.33, curve: .easeIn)
-        rotate.addAnimations {
-            let angle = down ? CGFloat.pi : 0
-            self.downloadsView.transform = CGAffineTransform(rotationAngle: angle)
-        }
-        rotate.startAnimation()
-    }
-
-    @objc private func handleBackPressed() {
-        webViewInterface?.goBack()
-        refreshNavigation()
-    }
-
-    @objc private func handleForwardPressed() {
-        webViewInterface?.goForward()
-        refreshNavigation()
-    }
-
-    @objc private func handleReloadPressed() {
-        webViewInterface?.reload()
-    }
-    
-    @objc private func handleActionsPressed() {
-        globalSettingsDelegate?.settingsDidPress(from: toolbarView, and: .zero)
-    }
-
     @objc private func handleShowOpenedTabsPressed() {
         coordinator?.showNext(.tabs)
     }
-
-    @objc private func handleDownloadsPressed() {
-        downloadsViewHidden.toggle() // should call didSet
-        downloadPanelDelegate?.didPressDownloads(to: downloadsViewHidden)
-    }
     
-    private func refreshNavigation() {
-        forwardButton.isEnabled = webViewInterface?.canGoForward ?? false
-        backButton.isEnabled = webViewInterface?.canGoBack ?? false
-    }
-    
-    private func updateToolbar(downloadsAvailable: Bool, actionsAvailable: Bool) {
-        var barItems = standardToolbarButtons
-        if actionsAvailable {
-            let space = UIBarButtonItem(barButtonSystemItem: .flexibleSpace, target: nil, action: nil)
-            barItems.append(space)
-            barItems.append(actionsButton)
+    private func handleDownloadsPressed() {
+        guard toolbarView.state.downloadsAvailable else {
+            return
         }
-        if downloadsAvailable {
-            let space = UIBarButtonItem(barButtonSystemItem: .flexibleSpace, target: nil, action: nil)
-            barItems.append(space)
-            barItems.append(downloadLinksButton)
-        }
-        toolbarView.setItems(barItems, animated: true)
+        toolbarView.state = .downloadsTapped
+        downloadPanelDelegate?.didPressDownloads(to: false)
     }
 }
 
 extension WebBrowserToolbarController: FullSiteNavigationComponent {
     func changeBackButton(to canGoBack: Bool) {
-        backButton.isEnabled = canGoBack
+        let prevState = toolbarView.state
+        toolbarView.state = .updateBackState(canGoBack)
+        toolbarView.state = prevState
     }
     
     func changeForwardButton(to canGoForward: Bool) {
-        forwardButton.isEnabled = canGoForward
+        let prevState = toolbarView.state
+        toolbarView.state = .updateForwardState(canGoForward)
+        toolbarView.state = prevState
     }
     
     func reloadNavigationElements(_ withSite: Bool, downloadsAvailable: Bool = false) {
-        // this will be useful when user will change current web view
-        backButton.isEnabled = webViewInterface?.canGoBack ?? false
-        forwardButton.isEnabled = webViewInterface?.canGoForward ?? false
-        reloadButton.isEnabled = withSite
-        updateToolbar(downloadsAvailable: downloadsAvailable, actionsAvailable: true)
-        downloadsViewHidden = !downloadsAvailable
-        enableDownloadsButton = downloadsAvailable
+        switch (withSite, downloadsAvailable) {
+        case (false, _):
+            toolbarView.state = .nothingToNavigate
+        case (true, false):
+            toolbarView.state = .readyForNavigation
+        case (true, true):
+            toolbarView.state = .readyForDownloads
+        }
     }
 
     var siteNavigator: WebViewNavigatable? {
@@ -271,25 +121,6 @@ extension WebBrowserToolbarController: FullSiteNavigationComponent {
         }
         set (newValue) {
             webViewInterface = newValue
-        }
-    }
-}
-
-extension CounterView: TabsObserver {
-    func update(with tabsCount: Int) {
-        self.digit = tabsCount
-    }
-}
-
-extension WebBrowserToolbarController: TabsObserver {
-    func tabDidSelect(index: Int, content: Tab.ContentType, identifier: UUID) {
-        switch content {
-        case .site:
-            updateToolbar(downloadsAvailable: false, actionsAvailable: true)
-        default:
-            // allow to show actions even for top sites view
-            // to be able for user to get to global settings
-            updateToolbar(downloadsAvailable: false, actionsAvailable: true)
         }
     }
 }
